Make `fetch_current_user` update the session
