+++ spec/rails_helper.rb
@@ -7,7 +7,13 @@ require "shoulda/matchers"
 
 Dir[Rails.root.join("spec/support/**/*.rb")].each { |file| require file }
 
+module Features
+  # Extend this module in spec/support/features/*.rb
+  include Features::OmniauthSignInHelper
+end
+
 RSpec.configure do |config|
+  config.include Features, type: :feature
   config.infer_base_class_for_anonymous_controllers = false
   config.infer_spec_type_from_file_location!
   config.use_transactional_fixtures = false
